requirements.txt
Flask>=3.0,<4
pypdf>=4.2,<5  # optional: PDF metadata extraction (scripts_doc_processing, metadata_compare)
requests>=2.32,<3  # optional: IP enrichment (asn_enrich.py)
python-docx>=1.1,<2  # optional: generate sample agreement (create_sample_agreement.py)
